refactor(orchestrator): extract tool-dispatch layer from executor_agent_loop
After the LLM-call layer moved out in a previous PR,
executor_agent_loop.go still mixed the Execute loop with tool
routing (built-in shell/git/file + MCP namespacing), sub-executor
invocation, and the gating predicates. Pull the tool-dispatch
cluster into a sibling so the next reader sees "execute loop +
events + step/artifact builders" stay together in one file and
"how a tool call becomes an Executor invocation" lives in
another.

  executor_agent_loop.go:       2112 → 1786 LOC
  executor_agent_loop_tools.go: new, 339 LOC

What moves to executor_agent_loop_tools.go

  Gating predicates
  - isGated
  - mcpServerPolicy
  - gatedToolsInTurn

  Tool routing
  - dispatchToolCall: maps an assistant tool_call into a built-in
    Executor (shell / git / file_edit / file_create / file_delete)
    or an MCP namespaced call.
  - isMCPToolName + dispatchMCPToolCall: the mcp__server__tool
    namespacing path, including the per-call telemetry emission
    around the upstream MCP handshake.

  Sub-executor invocation
  - runSubExecutor: wraps an Executor.Execute on a synthetic
    sub-task and shapes the result back into the tool-call return
    shape the LLM sees.

What stays in executor_agent_loop.go

  Construction (NewAgentLoopExecutor, SetMetrics,
  SetMCPHostFactory), the main Execute loop, the in-loop HTTP
  fetch tool (httpRequestTool — next 4d-http target), the
  recordMCPCallTelemetry helper (next 4d-telemetry target), the
  assistant-event emitters, the agentToolDefinitions catalog, and
  step/artifact builders.

The receiver type stays *AgentLoopExecutor everywhere; no
exported symbol moves; tests untouched.
executor_agent_loop_tools.go takes its own context/encoding-json/
errors/fmt/strings/time + internal/mcp/client (mcpclient alias) +
internal/telemetry + pkg/types.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
